Move deterministic loss generation method from manager to new utils module + refactoring changes
